+++ greatapi/routers/index.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter
 from fastapi import Depends
 
 from greatapi.db.models.user import User
-from greatapi.repositories.auth.jwt_token import get_current_user
+from greatapi.core.auth.jwt_token import get_current_user
 
 router = APIRouter(prefix='/index', tags=['Index'])
